refactor: remove old impl
